Finished Backups Update - Deleting backups - Restoring backups - es-ES updated - Translated some hardcoded keys. - Deleted mods from installations as I'll not keep track of those on the config to make it easier for users. They can download mods manually and they'll be shown automatically. - Fixed tasks not showing as finished. - Added a few more error logs just in case.

+++ src/ipc/workers/extractWorker.ts
@@ -3,7 +3,7 @@ import yauzl from "yauzl"
 import fse from "fs-extra"
 import path from "path"
 
-const { filePath, outputPath } = workerData
+const { filePath, outputPath, deleteZip } = workerData
 
 yauzl.open(filePath, { lazyEntries: true }, (err, zipfile) => {
   if (err) {
@@ -50,14 +50,16 @@ yauzl.open(filePath, { lazyEntries: true }, (err, zipfile) => {
           if (extractedCount === totalFiles) {
             zipfile.close()
 
-            fse.unlink(filePath, (err) => {
-              if (err) {
-                parentPort?.postMessage({ type: "error", message: `Error deleting ZIP file: ${err.message}` })
-                return
-              }
+            if (deleteZip) {
+              fse.unlink(filePath, (err) => {
+                if (err) {
+                  parentPort?.postMessage({ type: "error", message: `Error deleting ZIP file: ${err.message}` })
+                  return
+                }
+              })
+            }
 
-              parentPort?.postMessage({ type: "finished" })
-            })
+            parentPort?.postMessage({ type: "finished" })
           } else {
             zipfile.readEntry() // Leer siguiente entrada
           }
